refactor(App.js): 파일 구조 수정
