+++ src/App.js
@@ -1,13 +1,13 @@
 import { getValidatedCarNames, getValidatedAttemptCount } from "./inputHandler.js";
-import { simulateRace, displayWinners } from "./raceManager.js";
+import { simulateRace } from "./raceManager.js";
+import { displayWinners } from "./raceView.js";
 
 class App {
   async run() {
     const CARS = await getValidatedCarNames();
-
     const ATTEMPT_COUNT = await getValidatedAttemptCount();
-
     const RACE_RESULT = simulateRace(CARS, ATTEMPT_COUNT);
+
     displayWinners(RACE_RESULT);
   }
 }
